renamed files and added constants
